requirements.txt
Django==3.2.25
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
redis==5.0.7
celery==5.4.0
django-cors-headers==4.4.0
python-dotenv==1.0.1
elasticsearch==7.17.0
gunicorn==21.2.0

drf-spectacular==0.26.5

django-redis==5.4.0
django-filter==23.5